Use longs for event counts
In principle those might still wrap around, but at least it
less likely to happen.

+++ src/main/java/org/textensor/stochdiff/numeric/grid/DeterministicGridCalc.java
@@ -43,7 +43,7 @@ public class DeterministicGridCalc extends GridCalc {
     double[][] wktm1;
     double[][] wkC;
 
-    int event_count = 0;
+    long event_count = 0;
 
     public DeterministicGridCalc(int trial, SDRun sdm) {
         super(trial, sdm);
